requirements.txt
twitchAPI>=2.5.0
python-dotenv>=1.0.0
